patxh struttura

+++ croceitalia/src/main/java/org/cms/controller/croceitalia/StrutturaController.java
@@ -5,19 +5,13 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.cms.controller.SiteCmsController;
 import org.cms.controller.edit.EditCmsController;
-import org.cms.login.SoggettoUtente;
-import org.cms.login.Utente;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
-
-import it.asso.util.AssoException;
-
 
 @Controller
 public class StrutturaController extends EditCmsController {
@@ -26,17 +20,17 @@ public class StrutturaController extends EditCmsController {
 	 * 
 	 */
 	@Autowired(required = true)
-	protected StrutturaManager _strutturaManager;
-	
+	protected StrutturaManager			_strutturaManager;
+
 	@Autowired(required = true)
-	protected DocumentoTestataManager _documentoManager;
-	
+	protected DocumentoTestataManager	_documentoManager;
 
 	/**
 	 * @param request
 	 * @param response
 	 * @return
 	 */
+	@Override
 	@RequestMapping(value = "struttura/list", method = RequestMethod.GET)
 	protected ModelAndView list(HttpServletRequest request, HttpServletResponse response) {
 
@@ -78,7 +72,7 @@ public class StrutturaController extends EditCmsController {
 			return error(modelAndView, errore);
 		}
 	}
-	
+
 	@RequestMapping(value = "struttura/save")
 	protected ModelAndView save(HttpServletRequest request, HttpServletResponse response, Struttura str) {
 
@@ -96,17 +90,17 @@ public class StrutturaController extends EditCmsController {
 
 		String viewName = "croceitalia/struttura/list";
 		modelAndView.setViewName(viewName);
-		
+
 		return modelAndView;
 
 	}
-	
+
 	@RequestMapping(value = "struttura/create", method = RequestMethod.GET)
 	public ModelAndView create(HttpServletRequest request, HttpServletResponse response) {
 
 		ModelAndView modelAndView = getModelAndView(request);
 		try {
-			List<Tipologia_Struttura> tipologia= _strutturaManager.caricatipoStruttura();
+			List<Tipologia_Struttura> tipologia = _strutturaManager.caricatipoStruttura();
 			modelAndView.addObject("tipologia", tipologia);
 
 			modelAndView.setViewName("croceitalia/struttura/create");
@@ -118,37 +112,36 @@ public class StrutturaController extends EditCmsController {
 		}
 
 	}
-	
-	
+
 	@RequestMapping(value = "struttura/doUpdate", method = RequestMethod.POST)
 	public ModelAndView doUpdate(HttpServletRequest request, HttpServletResponse response, Struttura struttura) {
 
 		ModelAndView modelAndView = getModelAndView(request);
-		
+
 		try {
 			_strutturaManager.update(struttura);
 			request.setAttribute("esito", "ok");
 			String viewName = "forward:/edit/struttura/update/" + struttura.getId_struttura();
 			modelAndView.setViewName(viewName);
 			return modelAndView;
-			
+
 		} catch (Throwable errore) {
 			return error(modelAndView, errore);
 		}
 	}
-	
+
 	@RequestMapping(value = "struttura/update/{user_id}")
 	public ModelAndView update(HttpServletRequest request, HttpServletResponse response,
 			@PathVariable("user_id") String user_id) {
 
 		ModelAndView modelAndView = getModelAndView(request);
-		
+
 		try {
 			Struttura struttura = (Struttura) _strutturaManager.findById(user_id);
 			List<Tipologia_Struttura> tipo_struttura = _strutturaManager.caricatipoStruttura();
-			
+
 			modelAndView.addObject("tipo_struttura", tipo_struttura);
-			modelAndView.addObject("struttura", struttura);/*chiave valore*/
+			modelAndView.addObject("struttura", struttura);/* chiave valore */
 			modelAndView.setViewName("croceitalia/struttura/update");
 
 			return modelAndView;
@@ -157,58 +150,41 @@ public class StrutturaController extends EditCmsController {
 			return error(modelAndView, errore);
 		}
 	}
-	
-	
-//	@RequestMapping(value = "struttura/delete/{id}", method = RequestMethod.GET)
-//	public ModelAndView delete(HttpServletRequest request, HttpServletResponse response,
-//			@PathVariable("id") String id) {
-//
-//		ModelAndView modelAndView = getModelAndView(request);
-//
-//		try {
-//			List<Documento_Testata> struttura = _documentoManager.listaPerStruttura(id);
-//			String a="";
-//			for(int i=0;i<struttura.size();i++){
-//				a=a+"  "+ struttura.get(i).getNum_documento().toString();
-//			}
-//			if (struttura.size() > 0) {
-//				modelAndView.addObject("messaggio", "Cliente utilizzato, cancellazione impossibile. Il cliente è utilizzato nei documenti numero: "+a);
-//
-//			} else {
-//				_documentoManager.deleteById(id);
-//				modelAndView.addObject("messaggio", "Cancellazione effettuata correttamente");
-//			}
-//		} catch (Throwable errore) {
-//			return error(modelAndView, errore);
-//		}
-//
-//		List<Struttura> struttureList = _strutturaManager.caricaStruttura();
-//		modelAndView.addObject("Lista", struttureList);
-//
-//		String viewName = "croceitalia/struttura/list";
-//		modelAndView.setViewName(viewName);
-//
-//		return modelAndView;
-//
-//	}
-	
 
-	@RequestMapping(value = "struttura/delete/{id}/{pageId}", method = RequestMethod.GET)
-	public ModelAndView delete(HttpServletRequest request, HttpServletResponse response, @PathVariable("id") String id,
-			@PathVariable("pageId") String pageId) {
+	@RequestMapping(value = "struttura/delete/{id}", method = RequestMethod.GET)
+	public ModelAndView delete(HttpServletRequest request, HttpServletResponse response, @PathVariable("id") String id) {
 
 		ModelAndView modelAndView = getModelAndView(request);
-		try {
-			_strutturaManager.deleteById(id);
-			return pageAfterDelete(request, response, pageId);
 
+		try {
+
+			List<Documento_Testata> struttura = _documentoManager.listaPerStruttura(id);
+			String a = "";
+			for (int i = 0; i < struttura.size(); i++) {
+				a = a + "  " + struttura.get(i).getNum_documento().toString();
+			}
+			if (struttura.size() > 0) {
+				modelAndView.addObject("messaggio",
+						"Cliente utilizzato, cancellazione impossibile. Il cliente è utilizzato nei documenti numero: "
+								+ a);
+
+			} else {
+
+				_strutturaManager.deleteById(id);
+				modelAndView.addObject("messaggio", "Cancellazione effettuata correttamente");
+			}
 		} catch (Throwable errore) {
 			return error(modelAndView, errore);
 		}
 
+		List<Struttura> struttureList = _strutturaManager.caricaStruttura();
+		modelAndView.addObject("Lista", struttureList);
+
+		String viewName = "croceitalia/struttura/list";
+		modelAndView.setViewName(viewName);
+
+		return modelAndView;
+
 	}
-
-
-
 
 }
